+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change Log
 All notable changes to the "copyright-notice" extension will be documented in this file.
 
+## [1.1.1] - 2025-12-04
+### Added
+- `copyright-notice.autoRemoveEmojis` setting for automatic emoji removal on file save
+
 ## [1.1.0] - 2025-12-04
 ### Added
 - New "Remove All Emojis from File" command to clean emoji characters from code files
 - Comprehensive emoji detection covering all Unicode emoji ranges
+- `copyright-notice.autoRemoveEmojis` setting for automatic emoji removal on file save
 
 ## [1.0.1] - 2025-07-18
 ### Fixed